polish

+++ tests/deleting.rs
@@ -9,31 +9,38 @@ async fn deleting() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     
     let client: ScyllaClient = ScyllaClient::new(vec!["127.0.0.1"]).await?;
 
-    let delete_query_1: Result<QueryResult, QueryError> = client
+    let delete_query: Result<QueryResult, QueryError> = client
         .query("test_keyspace", "test_table")
         .delete()
-        .eq("age", 22)
+        .eq("user_id", 3)
+        .eq("age", 44)
         .execute()
         .await;
 
-    match &delete_query_1 {
-        Ok(query_result) => print_query_result("Query 1:", query_result),
+    match &delete_query {
+        Ok(query_result) => print_query_result("Query:", query_result),
         Err(e) => println!("Query 1failed: {:?}", e),
     }
+    Ok(())
+}
 
-    let delete_query_2: Result<QueryResult, QueryError> = client
-        .query("test", "test2_table")
+#[tokio::test]
+async fn in_list_deleting() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    
+    let client: ScyllaClient = ScyllaClient::new(vec!["127.0.0.1"]).await?;
+    let delete_query: Result<QueryResult, QueryError> = client
+        .query("test_keyspace", "test_table")
         .delete()
-        .in_list("id", &vec![22, 55])
+        .in_list("age", &vec![22, 55])
+        .in_list("user_id", &vec![11, 22])
         .execute()
         .await;
 
-    match &delete_query_2 {
-        Ok(query_result) => print_query_result("Query 2:", query_result),
-        Err(e) => println!("Query 2 failed: {:?}", e),
+    match &delete_query {
+        Ok(query_result) => print_query_result("Query:", query_result),
+        Err(e) => println!("Query failed: {:?}", e),
     }
 
 
     Ok(())
 }
-
